# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Proxy",
        "title": "API Proxy",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/repeat/{id}": {
            "post": {
                "description": "Resends the request and returns the result",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Repeat a request by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ParsedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Request not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to repeat request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requests": {
            "get": {
                "description": "Returns a list of all requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Get all requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Request"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch requests",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/requests/{id}": {
            "get": {
                "description": "Returns a request by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Get request by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Request"
                        }
                    },
                    "400": {
                        "description": "Invalid request ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Request not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ParsedRequest": {
            "type": "object",
            "properties": {
                "cookies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "get_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "post_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.ParsedResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Request": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "request": {
                    "$ref": "#/definitions/models.ParsedRequest"
                },
                "response": {
                    "$ref": "#/definitions/models.ParsedResponse"
                }
            }
        }
    }
}